+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.80
+
+- better ratings for disney title cards.
+
 ## 1.1.79
 
 - only pull ratings for visible elements, to reduce API calls